{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.0.1-alpha"
description = "{{cookiecutter.project_description}}"
authors = ["{{cookiecutter.project_author}}"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
sitri = "^0.5.5"
flask = "^1.1"
flask-sqlalchemy = "^2.4"
flask-migrate = "^2.5"
flask-jwt = "^0.3.2"
passlib = "^1.7"
python-slugify = "^3.0"
argon2-cffi = "^19.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.8"
pre-commit = "^1.18"
bump2version = "^0.5.11"
psycopg2-binary = "^2.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
